requirements.txt
fastapi
langchain_openai
langchain_core
langchain_community
tavily-python
python-dotenv
langgraph
pytest
ruff
langchain_experimental 
matplotlib
uvicorn 
pydantic 
python-jose[cryptography]
python-multipart
PyJWT
requests  
httpx
loguru